% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class_ColorMapper.R
\name{continuousMapper}
\alias{continuousMapper}
\title{Generate continuous value to color mapper}
\usage{
continuousMapper(colors = NA, range = NA, colo_of_NA = "grey")
}
\arguments{
\item{colors}{colors see description for details.}

\item{range}{The range of the colors to map, see description for details.}

\item{colo_of_NA}{color of NA}
}
\value{
colors
}
\description{
If you input a matrix, the range could be the max and min value of the matrix.

You can also input a numeric vector.

The return colors have the same dim with the input.

For the colors, you may need the \code{colorRampPalette} function to generate more colors.

For example: colorRampPalette(c("blue", "red"))( 50 ) , it will produce 50 colors that gradient from blue to red.
}
\examples{
values <- rnorm(10)
continuousMapper(colors = c("navy", "white", "firebrick3"), range = range(values))
continuousMapper(colors = colorRampPalette(c("blue", "red"))( 50 ), range = range(values))
}
