% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class_EdgeDrawer.R
\name{create_dege_drawer}
\alias{create_dege_drawer}
\title{Quick function to create the Edge drawer instance.}
\usage{
create_dege_drawer(arrowLength = 0.1, curvature = 1)
}
\value{
the drawer instance
}
\description{
Quick function to create the Edge drawer instance.
}
\examples{
create_dege_drawer()
}
