% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class_MembraneDNA_AlongLineDrawer.R
\name{create_DNA_drawer}
\alias{create_DNA_drawer}
\title{Quick function to create the DNA drawer instance.}
\usage{
create_DNA_drawer(dnaType = 1, numOfRepeats = 3, desired_height = 0.4)
}
\arguments{
\item{dnaType}{index of the DNA drawer type, from index 1}

\item{numOfRepeats}{the helix repeat, the greater the longer length}

\item{desired_height}{the height of the helix}
}
\value{
the drawer instance
}
\description{
Quick function to create the DNA drawer instance.
}
\examples{
create_DNA_drawer()
}
