% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class_LipidBilayerAlongLine.R
\name{create_lipidBilayer_drawer}
\alias{create_lipidBilayer_drawer}
\title{quick function to draw the instance.}
\usage{
create_lipidBilayer_drawer(lipidModelIndex = 1)
}
\arguments{
\item{lipidModelIndex}{the index of model}
}
\value{
the drawer instance
}
\description{
quick function to draw the instance.
}
\examples{
create_lipidBilayer_drawer()
}
