% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class_BioGraphicsObjs.R
\name{create_oval_node}
\alias{create_oval_node}
\title{Create the oval bio graphics node drawer.}
\usage{
create_oval_node(scaleWidth = 1, scaleHeight = 1, inner_extension_ratio = 0)
}
\arguments{
\item{scaleWidth}{scaler for width}

\item{scaleHeight}{scaler for height}

\item{inner_extension_ratio}{try this argument yourself}
}
\value{
the instance
}
\description{
Create the oval bio graphics node drawer.
}
\examples{
create_oval_node()
}
