% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class_Geometric.R
\name{create_rectangle}
\alias{create_rectangle}
\title{Create the Rectangle object.}
\usage{
create_rectangle(x = 0, y = 0, w = 1, h = 1)
}
\arguments{
\item{x}{left bottom point x}

\item{y}{left bottom point y}

\item{w}{width}

\item{h}{height}
}
\value{
The Rectangle object
}
\description{
Create the Rectangle object.
}
\examples{
create_rectangle(1,1,1,1)
}
