% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class_BioGraphicsObjs.R
\name{create_round_node}
\alias{create_round_node}
\title{Create the round bio graphics node drawer.}
\usage{
create_round_node(inner_extension_ratio = 0)
}
\arguments{
\item{inner_extension_ratio}{try this argument yourself}
}
\value{
the instance
}
\description{
Create the round bio graphics node drawer.
}
\examples{
create_round_node()
}
