% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class_Geometric.R
\name{create_round_rectangle_xyCoords}
\alias{create_round_rectangle_xyCoords}
\title{create a round rectangle xy coords, with the correct sequence order.}
\usage{
create_round_rectangle_xyCoords(
  x,
  y,
  w,
  h,
  r,
  space,
  angle_for_roundCorner = 5
)
}
\arguments{
\item{x}{left bottom x}

\item{y}{left bottom y}

\item{w}{width}

\item{h}{height}

\item{r}{radius of the circle}

\item{space}{the space of the two points}

\item{angle_for_roundCorner}{the angle in the round corner, in degree, default is 5}
}
\value{
a desired 2xn matrix
}
\description{
create a round rectangle xy coords, with the correct sequence order.
}
\examples{
This is a beautiful
coords <- create_round_rectangle_xyCoords(x = 1, y = 1, w = 5, h = 4, r = 1, space = 0.1,angle_for_roundCorner = 6)
}
