% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class_BioGraphicsObjs.R
\name{create_round_rectangular_node}
\alias{create_round_rectangular_node}
\title{Create the round rectangular bio graphics node drawer.}
\usage{
create_round_rectangular_node(
  width = 1,
  height = 1,
  r_ratio = 0.2,
  inner_extension_ratio = 0
)
}
\arguments{
\item{width}{the width with inchs}

\item{height}{the height with inchs}

\item{r_ratio}{the ratio of the round corner}

\item{inner_extension_ratio}{try this argument yourself}
}
\value{
the instance
}
\description{
Create the round rectangular bio graphics node drawer.
}
\examples{
create_round_rectangular_node()
}
