% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class_MembraneDNA_AlongLineDrawer.R
\name{create_roundRectBilayer_drawer}
\alias{create_roundRectBilayer_drawer}
\title{quick function to create the instance.}
\usage{
create_roundRectBilayer_drawer(
  numberOfIntervals = 10,
  percentageOfRectangle = 0.7
)
}
\arguments{
\item{numberOfIntervals}{intervals}

\item{percentageOfRectangle}{percentage of the round rectangle}
}
\value{
the drawer instance
}
\description{
quick function to create the instance.
}
\examples{
create_roundRectBilayer_drawer()
}
