% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class_BioGraphicsObjs.R
\name{create_selfContained_simple_bioGraphicsNode}
\alias{create_selfContained_simple_bioGraphicsNode}
\title{Create the object with the self contained simple bio graphics node.
This node is only has one shape x y coordinates.}
\usage{
create_selfContained_simple_bioGraphicsNode(name)
}
\arguments{
\item{name}{the name of the supported free shape.}
}
\value{
A BioGraphicNode instance.
}
\description{
use \code{basic_bioGraphics_templates |> names()} to see the supported name.
}
\examples{
create_selfContained_simple_bioGraphicsNode('LRP_1')
}
