% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class_TreeStructureInR6.R
\name{create_treeObj}
\alias{create_treeObj}
\title{Create the GraphicTree container for the rootNode.
It will also assign the number of nodes (member property of the node) for each node.}
\usage{
create_treeObj(rootNode)
}
\arguments{
\item{rootNode}{the root node of the tree}
}
\value{
the tree object instance
}
\description{
You may confused that the Tree container is redundant, for we could represent the tree just using the root node.
While some situations, we need the tree object to record the tree-level
}
\examples{
create_treeObj(rootNode)
}
