% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visuliza_RegulonAndCoexpression.R
\name{display_TF_targets_coexpress}
\alias{display_TF_targets_coexpress}
\title{Display the TF targets and co expressed genes.}
\usage{
display_TF_targets_coexpress(
  data = list(TF = "TF1", targets = paste0("g", c(4, 1, 9, 8, 3, 2, 6, 7, 5)),
    coexpression = c(paste0("g", 1:7), paste0("k", 1:3)))
)
}
\arguments{
\item{data:}{A list with TF, targets and coexpression. The contents are char vector.}
}
\description{
The input data list needs three elements, the name is fiexed: TF, targets and coexpression.
}
\examples{
display_TF_targets_coexpress()

#######
targets = c("")
}
