% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class_TreeStructureInR6.R
\name{displayTheTree}
\alias{displayTheTree}
\title{Display the tree by the simplest approach, that is , a table-like statement.}
\usage{
displayTheTree(node, fun = NULL)
}
\arguments{
\item{node:}{a node, may not be the rootNode.}

\item{fun:}{function(node), user defined fucntion for customized display.}
}
\value{
nothing to return
}
\description{
\if{html}{\out{
<pre>
For example, you have a tree like:
      |-------1
      |
   ---1.2
      |
      |---------2

The format is :
parent node    children number    child id

So, you get:
Current 1.2    2    1,2
Current 1  is leaf.
Current 2  is leaf.
</pre>
}}
}
\examples{
displayTheTree(rootNode)
displayTheTree(rootNode, function(x){
    cat("Hi ", x$xAxis_or_radius, "\n")
})
}
