% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class_Geometric.R
\name{distance}
\alias{distance}
\title{Get distance from two points.}
\usage{
distance(x0, y0, x1, y1)
}
\arguments{
\item{x0}{x of point1}

\item{y0}{y of point1}

\item{x1}{x of point2}

\item{y1}{y of point2}
}
\value{
distance
}
\description{
Get distance from two points.
}
\examples{
distance(0,0,1,1)
}
