% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class_MembraneDNA_AlongLineDrawer.R
\name{DNA_DoubleHelixDrawer}
\alias{DNA_DoubleHelixDrawer}
\title{The DNA double delix drawer, it can draw DNA along the curve.}
\description{
The draw of the delix is repeated, so it should has a drawer class.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{numOfRepeats}}{the number of repeats, default is 8}

\item{\code{listOfRibbonGpars}}{The graphics par of the round rectangle}

\item{\code{listOfHydrogenBondColors}}{The colors of the hydrogen bonds, the number should be 1,2,4,8}

\item{\code{dna_template_instance}}{The DNA template instance to paint.
It includes the graphics points of the HydrogenBbond
And the graphics par of the Ribbons}

\item{\code{desired_height}}{The height in inch's}

\item{\code{painting_parameters}}{store for the parameters for painting. (for more complex functions it can be used)}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DNA_DoubleHelixDrawer-draw_me}{\code{DNA_DoubleHelixDrawer$draw_me()}}
\item \href{#method-DNA_DoubleHelixDrawer-clone}{\code{DNA_DoubleHelixDrawer$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DNA_DoubleHelixDrawer-draw_me"></a>}}
\if{latex}{\out{\hypertarget{method-DNA_DoubleHelixDrawer-draw_me}{}}}
\subsection{Method \code{draw_me()}}{
Input the point of the location and it will draw the double helix


It will handle the angle automatically. Note the denser the dots, the more they are drawn
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DNA_DoubleHelixDrawer$draw_me(x, y)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{x location fro drawing, in inchs}

\item{\code{y}}{y location fro drawing, in inchs}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DNA_DoubleHelixDrawer-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DNA_DoubleHelixDrawer-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DNA_DoubleHelixDrawer$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
