% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{DNA_templates}
\alias{DNA_templates}
\title{Create the DNA double helix templates
20231103 updated}
\format{
a list in list structure
}
\source{
manually created my developers.
}
\usage{
DNA_templates
}
\description{
Create the DNA double helix templates
20231103 updated
}
\keyword{datasets}
