% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class_AffineTransformer.R
\name{do_affine_transformation}
\alias{do_affine_transformation}
\title{The core process to do the affine transformation.}
\usage{
do_affine_transformation(mat, transform_mat)
}
\arguments{
\item{mat}{the matrix that need to be transformed. The format is defined in the details.}

\item{transform_mat}{the transform matrix, always should be a 3 x 3 diminsion.}
}
\value{
the transformed matrix. the row number is 2, not three.
}
\description{
The algorithm is referenced from https://blog.csdn.net/u011681952/article/details/98942207

Be caution of the location of matrix and transform matrix.
}
\details{
The input matrix is defined as a 2 * n dimension.

For example, if you have 3 points, you should input a 2 x 3 matrix.

The output matrix is a 3 * n dimension matrix.
}
\examples{

mat <- rbind(c(1,0,1),c(2,1,0))
# we config a
transform_mat_tempelate <- matrix(c(1,0,0,0,1,0,0,0,1), byrow = T, ncol = 3)
transform_mat <- transform_mat_tempelate
transform_mat[1:2, 3] <- c(2,2)
# to do the transformation
after_mat <- do_affine_transformation(mat,transform_mat)

## you can visulize the results
library(grid)
grid.newpage()
grid.lines(x =mat[1,], y = mat[2,], default.units = 'in' )
after_mat <- do_affine_transformation(mat,transform_mat)
grid.lines(x =after_mat[1,], y = after_mat[2,], default.units = 'in' )
}
