% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class_AffineTransformer.R
\name{do_rotate_affine}
\alias{do_rotate_affine}
\title{The core process to do the rotate affine transformation.}
\usage{
do_rotate_affine(mat, theta = 0)
}
\arguments{
\item{mat}{the matrix that need to be transformed}

\item{theta}{the rotate angle, in radian}
}
\value{
the transformed matrix. the row number is 2, not three.
}
\description{
The core process to do the rotate affine transformation.
}
\examples{
mat <- rbind(c(1,0,1),c(2,1,0))
do_rotate_affine(mat, 1,1)

}
