% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class_AffineTransformer.R
\name{do_scale_affine}
\alias{do_scale_affine}
\title{The core process to do the scale affine transformation.}
\usage{
do_scale_affine(mat, scaleWidth = 1, scaleHeight = 1)
}
\arguments{
\item{mat}{the matrix that need to be transformed}

\item{scaleWidth}{scale in width}

\item{scaleHeight}{scale in height}
}
\value{
the transformed matrix. the row number is 2, not three.
}
\description{
The core process to do the scale affine transformation.
}
\examples{
mat <- rbind(c(1,0,1),c(2,1,0))
do_scale_affine(mat, 1,1)

}
