% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class_AffineTransformer.R
\name{do_translate_affine}
\alias{do_translate_affine}
\title{The core process to do the translate affine transformation.}
\usage{
do_translate_affine(mat, moveX = 1, moveY = 1)
}
\arguments{
\item{mat}{the matrix that need to be transformed}

\item{moveX}{the x move distance}

\item{moveY}{the y move distance}
}
\value{
the transformed matrix. the row number is 2, not three.
}
\description{
The core process to do the translate affine transformation.
}
\examples{
mat <- rbind(c(1,0,1),c(2,1,0))
do_translate_affine(mat, 1,1)

}
