% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class_GraphicElementDrawer.R
\name{draw_2xn_matrix_points}
\alias{draw_2xn_matrix_points}
\title{Draw the 2xn matrix in points}
\usage{
draw_2xn_matrix_points(..., default.units = "in")
}
\arguments{
\item{...}{the 2xn matrix list}

\item{default.units}{default is in}
}
\value{
null
}
\description{
Draw the 2xn matrix in points
}
\examples{
draw_2xn_matrix_points(rbind(1:5,rep(3,5)))
}
