% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Util_graphic_common.R
\name{draw_arc}
\alias{draw_arc}
\title{Draw the arc of a circle.}
\usage{
draw_arc(
  startDegree,
  endDegree,
  radius,
  xCenter = 0.5,
  yCenter = 0.5,
  default_unit = "in"
)
}
\arguments{
\item{startDegree}{the degree where to start}

\item{endDegree}{the degree where to end}

\item{radius}{the radius of the circle}

\item{xCenter}{the xCenter coordinate point}

\item{yCenter}{the yCenter coordinate point}

\item{default_unit}{coordinate unit}
}
\description{
The mechanism is that you calculate the (x,y) location of the each points.
And than draw it.
}
\examples{
draw_arc(startDegree = 0, endDegree = 68,radius = 1, xCenter = 2,yCenter = 2)
}
