% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class_GraphicElementDrawer.R
\name{draw_bezier_fourPoints}
\alias{draw_bezier_fourPoints}
\title{Draw the bezier curve for the four points.}
\usage{
draw_bezier_fourPoints(
  xx = c(0.2, 0.5, 0.3, 0) + 0.2,
  yy = c(1, 0, 0, 1) + 0.2,
  default_unit = "in",
  ...
)
}
\arguments{
\item{...}{}

\item{x:}{x axis are: (1,2,3,4)}

\item{y:}{the order same as y}

\item{default_unit:}{the coordinate unit}
}
\description{
The location of four points are:\if{html}{\out{
<pre>
4    1
  3    2
</pre>
}}
}
\examples{
draw_bezier_fourPoints()
}
