% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class_GraphicElementDrawer.R
\name{draw_bezier_twoPoints}
\alias{draw_bezier_twoPoints}
\title{Draw_bezier from two points.}
\usage{
draw_bezier_twoPoints(x0, y0, x1, y1, default_unit = "in", vertical = T, ...)
}
\arguments{
\item{x0}{x0}

\item{y0}{y0}

\item{x1}{x1}

\item{y1}{y1}

\item{default_unit}{coordinate unit}

\item{...}{transfer to the grob function}

\item{vertical:}{otherwise horizontal}
}
\description{
Draw_bezier from two points.
}
\examples{
draw_bezier_twoPoints()
}
