% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class_GraphicElementDrawer.R
\name{draw_lineArrow_accordingTo_grobAndPoints}
\alias{draw_lineArrow_accordingTo_grobAndPoints}
\title{draw_lineArrow_accordingTo_grobAndPoints}
\usage{
draw_lineArrow_accordingTo_grobAndPoints(
  node1,
  node2,
  x1,
  y1,
  x2,
  y2,
  arrow = NULL,
  ...
)
}
\arguments{
\item{node1}{grob 1}

\item{node2}{grob 2}

\item{x1}{x1}

\item{y1}{y1}

\item{x2}{x2}

\item{y2}{y2}

\item{arrow}{the arrow instance}

\item{...}{arguments pass to the grid.lines() function}
}
\description{
draw_lineArrow_accordingTo_grobAndPoints
}
\examples{
node1 <- get_global_bioGraphics_nodes_list()[['ligand']]
node2 <- get_global_bioGraphics_nodes_list()[['receptor']]
ligand$xAxis_or_radius, receptor$xAxis_or_radius
ligand$yAxis_or_angle, receptor$yAxis_or_angle
draw_lineArrow_accordingTo_grobAndPoints(node1,node2,x1,y1,x2,y2)
}
