% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class_GraphicElementDrawer.R
\name{draw_sector}
\alias{draw_sector}
\title{Draw a sector.}
\usage{
draw_sector(
  xCenter = 0.5,
  yCenter = 0.5,
  startDegree = 0,
  endDegree = 30,
  circle_out_radius = 1,
  circle_inner_radius = 1.1,
  fill_colo = "#00FFB0",
  border_colo = NA,
  default_unit = "in",
  ...
)
}
\arguments{
\item{xCenter}{the xCenter coordinate point}

\item{yCenter}{the yCenter coordinate point}

\item{startDegree}{start degree}

\item{endDegree}{end degree}

\item{circle_out_radius}{outter circle radius}

\item{circle_inner_radius}{inner circle radius}

\item{fill_colo}{filled color}

\item{border_colo}{border color}

\item{default_unit}{coordinate unit}

\item{...}{}
}
\value{
the polygon grob
}
\description{
What? you do not know the sector? Try to google it.
In order not to get the odd results. The function set the default.units = "in".
}
\examples{
draw_sector() # give result with default values
}
