% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class_EdgeDrawer.R
\name{Edge_attachedWithArrowDrawer}
\alias{Edge_attachedWithArrowDrawer}
\title{The edge drawer, it can draw edge with or without the arrow.}
\description{
In most cases, the edge needs to be draw-ed and with an arrow.
Now I model this process by create the class here.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{gpar_shape}}{the gpar for the shape}

\item{\code{gpar_text}}{the gpar for the text}

\item{\code{arrow_instance}}{the arrow instance.}

\item{\code{text_position}}{hjust and vjust for textGrob}

\item{\code{xyPoints}}{store the xy points for drawing}

\item{\code{label}}{the label of the edge}

\item{\code{curvature}}{control the angle of the lines, -1 and 1 are recommended.
initialize the function to get the default parameters.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Edge_attachedWithArrowDrawer-new}{\code{Edge_attachedWithArrowDrawer$new()}}
\item \href{#method-Edge_attachedWithArrowDrawer-draw_me}{\code{Edge_attachedWithArrowDrawer$draw_me()}}
\item \href{#method-Edge_attachedWithArrowDrawer-draw_grill_edge}{\code{Edge_attachedWithArrowDrawer$draw_grill_edge()}}
\item \href{#method-Edge_attachedWithArrowDrawer-draw_rectangular_edge}{\code{Edge_attachedWithArrowDrawer$draw_rectangular_edge()}}
\item \href{#method-Edge_attachedWithArrowDrawer-clone}{\code{Edge_attachedWithArrowDrawer$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Edge_attachedWithArrowDrawer-new"></a>}}
\if{latex}{\out{\hypertarget{method-Edge_attachedWithArrowDrawer-new}{}}}
\subsection{Method \code{new()}}{
In order to initialize the default class properties.

Draw the arrow according to the input points.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Edge_attachedWithArrowDrawer$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Edge_attachedWithArrowDrawer-draw_me"></a>}}
\if{latex}{\out{\hypertarget{method-Edge_attachedWithArrowDrawer-draw_me}{}}}
\subsection{Method \code{draw_me()}}{
Input the point of the location and it will draw the arrow.

Draw the edge either horizontal or vertical.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Edge_attachedWithArrowDrawer$draw_me(x1, y1, x2, y2)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x1}}{x of point1}

\item{\code{y1}}{y of point1}

\item{\code{x2}}{x of point2}

\item{\code{y2}}{y of point2}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Edge_attachedWithArrowDrawer-draw_grill_edge"></a>}}
\if{latex}{\out{\hypertarget{method-Edge_attachedWithArrowDrawer-draw_grill_edge}{}}}
\subsection{Method \code{draw_grill_edge()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Edge_attachedWithArrowDrawer$draw_grill_edge(x1, y1, x2, y2)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x1}}{x of point1}

\item{\code{y1}}{y of point1}

\item{\code{x2}}{x of point2}

\item{\code{y2}}{y of point2
Draw the edge with diagonal line manner.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Edge_attachedWithArrowDrawer-draw_rectangular_edge"></a>}}
\if{latex}{\out{\hypertarget{method-Edge_attachedWithArrowDrawer-draw_rectangular_edge}{}}}
\subsection{Method \code{draw_rectangular_edge()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Edge_attachedWithArrowDrawer$draw_rectangular_edge(x1, y1, x2, y2)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x1}}{x of point1}

\item{\code{y1}}{y of point1}

\item{\code{x2}}{x of point2}

\item{\code{y2}}{y of point2}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Edge_attachedWithArrowDrawer-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Edge_attachedWithArrowDrawer-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Edge_attachedWithArrowDrawer$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
