% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class_API.R
\name{edges}
\alias{edges}
\title{Assign the edges component in this function.}
\usage{
edges(fix_indicator_df = NULL, free_indicator_df = NULL)
}
\arguments{
\item{fix_indicator_df}{data.frame of fixed indicator, see \code{details} for format}

\item{free_indicator_df}{data.frame of free indicator, see \code{details} for format}
}
\value{
the \code{pDescriptor} instance
}
\description{
Assign the edges component in this function.
}
\details{
The \code{fix_indicator_df} should must contain: \code{from} and \code{to} two columns.
The \code{free_indicator_df} should must contain: \code{x1}, \code{y1} , \code{x2} and \code{y2} four columns.
The \code{curvature} and \code{arrowLen} are optional, default values are 0 and 0.2 inchs.
}
\examples{
## Add the free arrow
df <- rbind(
  c(4.6,1.45,5,1.45,-1,0.1),
  c(6.2,1.45,6.6,1.45,-1,0.1),
  c(2.5,3.1,2.9,2.7,-1,0.1),
  c(3.5, 2.3, 4.3, 2,-1,0.1)
)

colnames(df) <- c('x1','y1','x2','y2','curvature','arrowLen')

edges(free_indicator_df = df)

}
