% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class_Geometric.R
\name{fourPoints_to_calculate_bezier}
\alias{fourPoints_to_calculate_bezier}
\title{Quick function to calculate the bezier lines.}
\usage{
fourPoints_to_calculate_bezier(points, n_step_bezier = 40)
}
\arguments{
\item{points}{the Point instance}
}
\value{
the 2xn matrix of xy points
}
\description{
Quick function to calculate the bezier lines.
}
\examples{
Please see the [[soomthOut_twoSidesOf_quadrangle()]]
}
