% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class_ColorMapper.R
\name{generate_colors}
\alias{generate_colors}
\title{Generate color for the input x}
\usage{
generate_colors(x, values2colorMapper)
}
\arguments{
\item{x}{matrix or numeric vector}

\item{values2colorMapper}{}
}
\value{
colors
}
\description{
Generate color for the input x
}
\examples{
a <- discreteMapper(colors = c('red', 'blue'), break_points = c(1,2,3))
generate_colors(1:10,a)
}
