% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class_ColorMapper.R
\name{generate_equal_spersed_intervals}
\alias{generate_equal_spersed_intervals}
\title{Generate equal spersed intervals from x numeric values according to the number of intervals and center parameter.}
\usage{
generate_equal_spersed_intervals(x, numOfBreakPoints, center = F)
}
\arguments{
\item{x}{: numeric values, e.g. (1,1.2,3,0.5)}

\item{numOfBreakPoints}{: number of intervals (number of break points - 1)}

\item{center}{: should the break points, equally distribute on the two side of the 0.}
}
\value{
break_points: a vector consists of the break points, e.g. c(1,3,5,7)
}
\description{
Generate equal spersed intervals from x numeric values according to the number of intervals and center parameter.
}
\examples{
x = c(1,2,3,4,5,6,7), numOfIntervals = 3, center =F
generate_equal_spersed_intervals(x, numOfIntervals,center) will
return c(1,4,7) where 1,4,7 is the break points.
}
