% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Global_variables.R
\name{get_global_bioGraphics_nodes_list}
\alias{get_global_bioGraphics_nodes_list}
\title{get_global_bioGraphics_nodes_list}
\usage{
get_global_bioGraphics_nodes_list()
}
\value{
a list
}
\description{
get_global_bioGraphics_nodes_list
}
\examples{
get_global_bioGraphics_nodes_list() |> names()
}
