% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Global_variables.R
\name{get_global_pars}
\alias{get_global_pars}
\title{Get global parameters.}
\usage{
get_global_pars()
}
\value{
the global pars
}
\description{
Get global parameters.
}
\examples{
get_global_pars() |> names()
}
