% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Global_variables.R
\name{get_global_text_pars}
\alias{get_global_text_pars}
\title{Quick get the global text gpar}
\usage{
get_global_text_pars()
}
\value{
gpar class of text
}
\description{
Quick get the global text gpar
}
\examples{
get_global_text_pars()
}
