% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class_AffineTransformer.R
\name{get_rotate_transform_mat}
\alias{get_rotate_transform_mat}
\title{The low-level function for rotate transforming}
\usage{
get_rotate_transform_mat(theta = 0)
}
\arguments{
\item{theta}{the rotate angle, in radian}
}
\value{
the transforming matrix not the transformed matrix
}
\description{
The low-level function for rotate transforming
}
\examples{
get_scale_transform_mat(2,1)
}
