% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class_AffineTransformer.R
\name{get_scale_transform_mat}
\alias{get_scale_transform_mat}
\title{The low-level function for scale transforming}
\usage{
get_scale_transform_mat(scaleWidth = 1, scaleHeight = 1)
}
\arguments{
\item{scaleWidth}{scale in width}

\item{scaleHeight}{scale in height}
}
\value{
the transforming matrix not the transformed matrix
}
\description{
The low-level function for scale transforming
}
\examples{
get_scale_transform_mat(2,1)
}
