% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class_AffineTransformer.R
\name{get_translate_transform_mat}
\alias{get_translate_transform_mat}
\title{The low-level function for translate transforming}
\usage{
get_translate_transform_mat(moveX = 1, moveY = 1)
}
\arguments{
\item{moveX}{the x move distance}

\item{moveY}{the y move distance}
}
\value{
the transforming matrix not the transformed matrix
}
\description{
The low-level function for translate transforming
}
\examples{
get_translate_transform_mat(2,1)
}
