% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class_TreeStructureInR6.R
\name{getLongestDepth}
\alias{getLongestDepth}
\title{The the longest depth of a tree.}
\usage{
getLongestDepth(node)
}
\arguments{
\item{node}{: root node}
}
\value{
: the value of the longest depth.
}
\description{
For a tree, it has lots of lineages, from root to the leaf.
This function will find the longest depth from root to a specific leaf.
}
\examples{
getLongestDepth(rootNode)
}
