% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class_TreeStructureInR6.R
\name{GraphicTree}
\alias{GraphicTree}
\title{The \code{graphicTree} class to represent the tree.}
\description{
The \code{graphicTree} class to represent the tree.

The \code{graphicTree} class to represent the tree.
}
\details{
\if{html}{\out{
<pre>
* It most has a rootNode
* The longestDepth and numOfLeaves is employed for the layout computation
</pre>
}}
}
\examples{

## ------------------------------------------------
## Method `GraphicTree$assignAttributes`
## ------------------------------------------------

instance$assignAttributes()
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{rootNode}}{The root node of the tree.}

\item{\code{longestDepth}}{The longest depth from the leaf to the root node.}

\item{\code{numOfLeaves}}{The total number of leaf in the tree.}

\item{\code{isPropertiesAssigned}}{Whether the tree instance properties are assigned.}

\item{\code{startDegree}}{start degree of the circle}

\item{\code{extendDegree}}{extend degree of the circle}

\item{\code{innerRadius}}{inner radius of the circle}

\item{\code{outterRadius}}{outter radius of the circle}

\item{\code{direction}}{the direction of the circle}

\item{\code{xCenter}}{the xCenter of the circle}

\item{\code{yCenter}}{the yCenter of the circle
Initialize the instance by input the rootNode.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-GraphicTree-new}{\code{GraphicTree$new()}}
\item \href{#method-GraphicTree-assignAttributes}{\code{GraphicTree$assignAttributes()}}
\item \href{#method-GraphicTree-clone}{\code{GraphicTree$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GraphicTree-new"></a>}}
\if{latex}{\out{\hypertarget{method-GraphicTree-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GraphicTree$new(rootNode)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{rootNode}}{root of the tree
To assign the \code{numOfLeaves} and \code{longestDepth} properties.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GraphicTree-assignAttributes"></a>}}
\if{latex}{\out{\hypertarget{method-GraphicTree-assignAttributes}{}}}
\subsection{Method \code{assignAttributes()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GraphicTree$assignAttributes()}\if{html}{\out{</div>}}
}

\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{instance$assignAttributes()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GraphicTree-clone"></a>}}
\if{latex}{\out{\hypertarget{method-GraphicTree-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GraphicTree$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
