% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{lipid_models}
\alias{lipid_models}
\title{Create the lipid models, the model is stored in this.}
\format{
a list in list structure

To get the first lipid, you need to \code{lipid_models[[1]]}
}
\source{
manually created my developers.
}
\usage{
lipid_models
}
\description{
Create the lipid models, the model is stored in this.
}
\keyword{datasets}
