% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class_NodeMaker.R
\name{make_complex_fromNodes}
\alias{make_complex_fromNodes}
\title{Make complex graphic node from a plenty of nodes.}
\usage{
make_complex_fromNodes(bioNodes = get_global_bioGraphics_nodes_list())
}
\arguments{
\item{bioNodes}{is list of r package level grobs . @seealso \code{\link[=get_global_bioGraphics_nodes_list]{get_global_bioGraphics_nodes_list()}}}
}
\value{
list of [\link{basic_bioGraphics_templates}]
}
\description{
Make complex graphic node from a plenty of nodes.
}
\examples{
make_complex_fromNodes()
}
