% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class_PathwayAPI.R
\name{membrane}
\alias{membrane}
\title{Prepare the membrane instance parameters}
\usage{
membrane(
  shape = c("ellipse", "roundRect", "points"),
  x = 0.5,
  y = 0.5,
  w = 2,
  h = 1.5,
  xyLocation = NULL,
  phos_height = 0.15,
  phos_dist = 0.1
)
}
\arguments{
\item{shape}{'ellipse','roundRect', 'points' see details}

\item{x}{x of the start point}

\item{y}{y of the start point}

\item{w}{width of the shape}

\item{h}{height of the shape}

\item{xyLocation}{x,y locations of the shape, the format is 2xn matrix}

\item{phos_height}{height of the phospholipid model, in 'inchs' not apply the \code{pScaleCoord}}

\item{phos_dist}{the distance between two phospholipid}
}
\value{
the \code{pDescriptor} instance
}
\description{
A ggplot style to create the parameters related to the membrane.
}
\details{
Users can specify necessary parameters with two approaches.
The first way is input the \code{shape}, \code{x}, \code{y}, \code{w} and \code{h}.
Or by input the \code{xyLocation} of the shape, the format is same as the conventions: 2xn matrix.

In \code{shape} parameter, the
ellipse and round rectangle are the regular shape.

\code{points} means the users could directly import the x and y vector that contains the x/y locations.
In this case the w and h parameters are ignored.
}
\examples{
membrane(shape = 'ellipse', x = 1, y = 0.2, w = 3, h =2)
}
