% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class_API.R
\name{nodes}
\alias{nodes}
\title{The main pathway component expressed in here.}
\usage{
nodes(data, customizedDF = NULL)
}
\arguments{
\item{data}{the needed data.frame to draw the figure}

\item{customizedDF}{the customized elements in the data. see [\link{details}] for more information}
}
\value{
the \code{pDescriptor} instance
}
\description{
\if{html}{\out{
<pre>
   Feature information:
   label : the name of the node
   x     : the x axis of the node
   y     : the y axis of the node
   width : the width the node, it can be used to change the circle to the ellipse
   height: the height of the node, it can be used to change the square to rectangle
   inner_extension_ratio
         : the extension ratio to make the pretty visual effect
   drawLabel
         : whether draw the label, default T, if the label not NA, the label will be drawed
   text_position
         : the numeric two elements vector, values between 0 and 1
   gpar_shape
         : the graphics parameters of the shape, same as the [grid] package.
   gpar_text
         : the graphics parameters of the text, same as the [grid] package.
   shape : the shape can be 'ellipse', 'rect' or 'roundRect', which means ellipse, rectangle and round corner rectangle.
           the node can be a template of the self-contained icon shape. see [create_selfContained_simple_bioGraphicsNode()]
</pre>
}}
}
\details{
Note: the \code{data} must should contain the columns of \code{shape/x/y/width/height} and the names should match exactly.

If the \code{customizedDF} is non-null, the \code{label} should exist in the \code{data} and other features should be in the \code{customizedDF} data.frame.
}
\examples{
df <- rbind(
  list('Lipoxygenases',5,4,0.4,0.6,0.2,c(0.5, 1.1), 'roundRect'),
  list(label='PPARa-1',x=3.2,y=2.5,w=1,h=0.8,ratio=0,tp=c(0.5, 0.5),'ellipse'),
  list(label='PPARa-2',x=4.8,y=2,w=0.7,h=0.8*0.7,ratio=0,tp=c(0.5, 0.5),'ellipse'),
  list(label='RXR-1',x=5,y=1.8,w=0.5,h=0.4,ratio=0.2,tp=c(0.5, 0.5),'roundRect'),
  list(label='PPARy',x=6,y=2,w=0.7,h=0.8*0.7,ratio=0,tp=c(0.5, 0.5),'ellipse'),
  list(label='RXR-2',x=6.3,y=1.8,w=0.5,h=0.4,ratio=0.2,tp=c(0.5, 0.5),'roundRect'),
  list(label=NA,x=4.3,y=1.45,w=0.5,h=0.4,ratio=0.2,tp=c(0.5, 0.5),'roundRect'),
  list(label=NA,x=5.9,y=1.45,w=0.5,h=0.4,ratio=0.2,tp=c(0.5, 0.5),'roundRect')
)
colnames(df) <- c("label","x","y","width","height","inner_extension_ratio","text_position", "shape")

nodes(data = df)
}
