% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class_API.R
\name{nuclearDNA}
\alias{nuclearDNA}
\title{Prepare the nuclear DNA instance parameters}
\usage{
nuclearDNA(x, y, dnaType = 2, numOfRepeats = 3, desired_height = 0.35)
}
\arguments{
\item{x}{x of the start point}

\item{y}{y of the start point}

\item{dnaType}{choose the visual effects of the DNA, currently we support two types, 1 or 2}

\item{numOfRepeats}{the repeat of the DNA double helix motif}

\item{desired_height}{the height of the DNA}
}
\value{
the \code{pDescriptor} instance
}
\description{
Prepare the nuclear DNA instance parameters
}
\examples{
nuclearDNA(x = 3.5, y = 0.7)
}
