% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class_API.R
\name{nuclearEnvelop}
\alias{nuclearEnvelop}
\title{Prepare the nuclear envelop instance parameters}
\usage{
nuclearEnvelop(
  shape = c("ellipse", "roundRect"),
  x = 2,
  y = 2,
  w = 1,
  h = 1,
  numOfIntervals = 10,
  percentageOfRectangle = 0.7,
  phos_height = 0.06
)
}
\arguments{
\item{shape}{c('ellipse','roundRect'), current support two of them}

\item{x}{x of the start point}

\item{y}{y of the start point}

\item{w}{width of the shape}

\item{h}{height of the shape}

\item{numOfIntervals}{number of the intervals}

\item{percentageOfRectangle}{the percentage of the rectangle}

\item{phos_height}{the phosolipid model height, in 'inchs' not apply the \code{pScaleCoord}}
}
\value{
the \code{pDescriptor} instance
}
\description{
Prepare the nuclear envelop instance parameters
}
\examples{
nuclearEnvelop(x = 3.3, y = 0.5, w = 1, h = 0.6)
}
