% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class_Geometric.R
\docType{data}
\name{ONE_DEGREE_IN_RADIAN}
\alias{ONE_DEGREE_IN_RADIAN}
\title{Define the constont values}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
ONE_DEGREE_IN_RADIAN
}
\description{
Define the constont values
}
\keyword{datasets}
