% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class_Java_interaction.R
\name{open_locationPicker_javaPrograme}
\alias{open_locationPicker_javaPrograme}
\title{Open the GUI program written in the JAVA language.}
\usage{
open_locationPicker_javaPrograme()
}
\description{
The major purpose is to combine the merits of R and JAVA language.
and bridge the gap between scripted program and GUI software.
}
\details{
Why the \code{SkeletonMaker} exists?
The biggest reason is that the R can not do click click and click.

Technique:

Note: currently, we do not automatically reference the package \code{rJava} as the dependence in this package.

Users need to install in yourself. The installation process is very easy. Just run \verb{install.packages('rJava)}.

Any problems, please contact us
}
\examples{
open_locationPicker_javaPrograme()
}
