% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Extra_Visulize_RegulonWithCoexpression.R
\name{outbound_painter}
\alias{outbound_painter}
\title{Outbound painter}
\usage{
outbound_painter(
  ol = list(x_TF_top = 2, y_TF_top = 2.5, x_TF_left = 0.5, y_TF_left = 2, x_TF_right =
    2.5, y_TF_right = 2, x_coeNodes_leftMost = 0, y_coeNodes_leftMost = 0.5,
    x_coeNodes_1_afMost = 0.5, y_coeNodes_1_afMost = 0, x_coeNodes_last_afMost = 2.5,
    y_coeNodes_last_afMost = 0, x_coeNodes_rightMost = 3, y_coeNodes_rightMost = 0.5,
    circle_unit_r_value = 0.3),
  coexpresNodes_belongsTo_targets,
  default_unit = "in",
  buttomStyleBezier = T,
  half_unit_coexpression_node = 0.5
)
}
\arguments{
\item{ol}{: outlinePoints}

\item{default_unit}{}
}
\description{
Outbound painter
}
\examples{
outbound_painter()
}
