% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PathwayIlluminator-package.R
\docType{package}
\name{PathwayIlluminator-package}
\alias{PathwayIlluminator}
\alias{PathwayIlluminator-package}
\title{PathwayIlluminator: What the Package Does (One Line, Title Case)}
\description{
The beautifull pathway drawer.
}
\author{
\strong{Maintainer}: Dalang Yu \email{dalangyusci@gmail.com}

}
\keyword{internal}
