% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class_PathwayAPI.R
\name{pDescriptor}
\alias{pDescriptor}
\title{The parameter preparation of the painting process.}
\description{
Inspired by the \code{ggplot2} package, We organized the parameters with the '+' symbols.
}
\examples{

## ------------------------------------------------
## Method `pDescriptor$print`
## ------------------------------------------------

print()
# Note: this function may not need to be directly invoked.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{membrane}}{the cell membrane parameters.}

\item{\code{nuclearEnvelop}}{the nuclear envelop parameters.}

\item{\code{nuclearDNA}}{the nuclear DNA parameters.}

\item{\code{nodes}}{the nodes parameters.}

\item{\code{edges}}{the edges parameters.}

\item{\code{name}}{the name parameter. The title is stored here.}

\item{\code{gpar_title}}{the gpar of the title.}

\item{\code{scaleCoordinate}}{the coordinate re-scale.

pp$scaleCoordinate <- c(realWidth, realHeight);}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-pDescriptor-print}{\code{pDescriptor$print()}}
\item \href{#method-pDescriptor-draw_membrane}{\code{pDescriptor$draw_membrane()}}
\item \href{#method-pDescriptor-draw_nuclearDNA}{\code{pDescriptor$draw_nuclearDNA()}}
\item \href{#method-pDescriptor-draw_edges}{\code{pDescriptor$draw_edges()}}
\item \href{#method-pDescriptor-draw_nodes}{\code{pDescriptor$draw_nodes()}}
\item \href{#method-pDescriptor-draw_nuclearEnvelop}{\code{pDescriptor$draw_nuclearEnvelop()}}
\item \href{#method-pDescriptor-draw_title}{\code{pDescriptor$draw_title()}}
\item \href{#method-pDescriptor-clone}{\code{pDescriptor$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-pDescriptor-print"></a>}}
\if{latex}{\out{\hypertarget{method-pDescriptor-print}{}}}
\subsection{Method \code{print()}}{
Workhouse function here.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pDescriptor$print()}\if{html}{\out{</div>}}
}

\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{print()
# Note: this function may not need to be directly invoked.
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-pDescriptor-draw_membrane"></a>}}
\if{latex}{\out{\hypertarget{method-pDescriptor-draw_membrane}{}}}
\subsection{Method \code{draw_membrane()}}{
Draw the membrane of the figure

We separate the drawing process, this is the membrane drawing process
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pDescriptor$draw_membrane(membrane)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{membrane}}{the list of the parameters}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-pDescriptor-draw_nuclearDNA"></a>}}
\if{latex}{\out{\hypertarget{method-pDescriptor-draw_nuclearDNA}{}}}
\subsection{Method \code{draw_nuclearDNA()}}{
Draw the nuclear DNA of the figure

We separate the drawing process, this is the DNA drawing process
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pDescriptor$draw_nuclearDNA(nuclearDNA)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{nuclearDNA}}{the list of input parameters}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-pDescriptor-draw_edges"></a>}}
\if{latex}{\out{\hypertarget{method-pDescriptor-draw_edges}{}}}
\subsection{Method \code{draw_edges()}}{
Draw the edges of the figure

We separate the drawing process, this is the edges drawing process

Draw the nodes of the figure
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pDescriptor$draw_edges()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-pDescriptor-draw_nodes"></a>}}
\if{latex}{\out{\hypertarget{method-pDescriptor-draw_nodes}{}}}
\subsection{Method \code{draw_nodes()}}{
We separate the drawing process, this is the nodes drawing process
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pDescriptor$draw_nodes()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-pDescriptor-draw_nuclearEnvelop"></a>}}
\if{latex}{\out{\hypertarget{method-pDescriptor-draw_nuclearEnvelop}{}}}
\subsection{Method \code{draw_nuclearEnvelop()}}{
Draw the nuclear envelop of the figure

We separate the drawing process, this is the nuclear envelop drawing process

Draw the title of the figure
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pDescriptor$draw_nuclearEnvelop(nuclearEnvelop)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{nuclearEnvelop}}{the list of the parameters}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-pDescriptor-draw_title"></a>}}
\if{latex}{\out{\hypertarget{method-pDescriptor-draw_title}{}}}
\subsection{Method \code{draw_title()}}{
We separate the drawing process, this is the title drawing process
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pDescriptor$draw_title()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-pDescriptor-clone"></a>}}
\if{latex}{\out{\hypertarget{method-pDescriptor-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pDescriptor$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
