% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class_LipidBilayerAlongLine.R
\name{PhospholipidBilayerDrawer}
\alias{PhospholipidBilayerDrawer}
\title{The phospholipid model drawer, it can draw phospholipid along the curve.}
\description{
The draw of the phospholipid is repeated, so it should has a drawer class.
}
\examples{

## ------------------------------------------------
## Method `PhospholipidBilayerDrawer$draw_one_unit_lipid`
## ------------------------------------------------

x <- seq(from = 1, to = 6, by = 0.6)
# the simplest form of the line
y <- x ^ (1 /3) + 1
# y <- x
grid.newpage()
painter <- Phospholipid_drawer$new()
painter$draw_lipid_along_curve(x = x, y = y)
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{shape_scaling_factor}}{The scaling parameter, bigger value for big lipid in your eyes.}

\item{\code{current_lipid_model}}{The model of the lipid}

\item{\code{lipid_head_par}}{The graphics par}

\item{\code{desired_hight}}{The height in inch's}

\item{\code{draw_circle_head}}{most draw head as a circle in a curve}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PhospholipidBilayerDrawer-draw_lipid_along_curve}{\code{PhospholipidBilayerDrawer$draw_lipid_along_curve()}}
\item \href{#method-PhospholipidBilayerDrawer-draw_one_unit_lipid}{\code{PhospholipidBilayerDrawer$draw_one_unit_lipid()}}
\item \href{#method-PhospholipidBilayerDrawer-clone}{\code{PhospholipidBilayerDrawer$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PhospholipidBilayerDrawer-draw_lipid_along_curve"></a>}}
\if{latex}{\out{\hypertarget{method-PhospholipidBilayerDrawer-draw_lipid_along_curve}{}}}
\subsection{Method \code{draw_lipid_along_curve()}}{
Input the curve line on the x,y parameters, the lipid will correctly draw along the line.


It will handle the angle automatically. Note the denser the dots, the more they are drawn

Draw one unit lipid model
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PhospholipidBilayerDrawer$draw_lipid_along_curve(lineX, lineY, lastDraw = F)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{lineX}}{x values}

\item{\code{lineY}}{y values}

\item{\code{lastDraw}}{should draw last model, default F}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PhospholipidBilayerDrawer-draw_one_unit_lipid"></a>}}
\if{latex}{\out{\hypertarget{method-PhospholipidBilayerDrawer-draw_one_unit_lipid}{}}}
\subsection{Method \code{draw_one_unit_lipid()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PhospholipidBilayerDrawer$draw_one_unit_lipid(
  scaleWidth = 1,
  scaleHeight = 1,
  theta,
  moveX,
  moveY
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{scaleWidth}}{the width scale factor}

\item{\code{scaleHeight}}{the height scale factor}

\item{\code{theta}}{angle in radian}

\item{\code{moveX}}{the moved x distance}

\item{\code{moveY}}{the moved y distance}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{x <- seq(from = 1, to = 6, by = 0.6)
# the simplest form of the line
y <- x ^ (1 /3) + 1
# y <- x
grid.newpage()
painter <- Phospholipid_drawer$new()
painter$draw_lipid_along_curve(x = x, y = y)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PhospholipidBilayerDrawer-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PhospholipidBilayerDrawer-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PhospholipidBilayerDrawer$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
