% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class_PathwayAPI.R
\name{+.pDescriptor}
\alias{+.pDescriptor}
\title{The main plus function to combine the parameters.}
\usage{
\method{+}{pDescriptor}(e1, e2)
}
\arguments{
\item{e1}{the \code{pDescriptor} instance}

\item{e2}{the \code{pDescriptor} instance}
}
\value{
the \code{pDescriptor} instance
}
\description{
The returned instance is the later \code{pDescriptor} isntance.
}
\examples{
p1 <- pDescriptor$new()
p1$name = 'p1'

p2 <- pDescriptor$new()
p2$name = 'p2'

p1 + p2
}
