% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class_TreeStructureInR6.R
\name{preOrder_traversal}
\alias{preOrder_traversal}
\title{A template to traversal all nodes.}
\usage{
preOrder_traversal(node, fun)
}
\arguments{
\item{fun}{: the function to input, the input parameter is only the node.}

\item{node:}{any tree node.}
}
\value{
nothing to return
}
\description{
A template to traversal all nodes.
}
\examples{
printSomeThing <- function(node){
   cat("Hi I am print ", node$id, "\n")
}
preOrder_traversal(rootNode, printSomeThing)
}
