% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class_TreeStructureInR6.R
\name{process_createNodes}
\alias{process_createNodes}
\title{Create a quick simulated tree.}
\usage{
process_createNodes(depth = 1, numOfChildren = 2)
}
\arguments{
\item{depth}{the depth of the tree, since the tree is balanced, so the tree look same. @seealso \code{\link[=getLongestDepth]{getLongestDepth()}}}

\item{numOfChildren}{n-fork tree}
}
\value{
the tree root node.
}
\description{
The tree is balanced and full children size. \if{html}{\out{<br>}}
What is this? Try yourself.
}
\examples{

process_createNodes(depth = 2, numOfChildren = 3)
}
