% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class_TreeStructureInR6.R
\name{process_dendrogram2intuitiveTree}
\alias{process_dendrogram2intuitiveTree}
\title{A process for dendrogram object  to GraphicTree instance.}
\usage{
process_dendrogram2intuitiveTree(dendrogram)
}
\arguments{
\item{dendrogram}{}
}
\value{
a GraphicTree instance
}
\description{
Convert the dendrogram object to the GraphicTree object.
}
\examples{
hc <- hclust(dist(USArrests), "ave")
dendrogram <- as.dendrogram(hc)
intuitiveTree <- process_dendrogram2intuitiveTree(dendrogram)
displayTheTree(intuitiveTree$rootNode)
}
