% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class_AffineTransformer.R
\name{produce_model_coordinate_points}
\alias{produce_model_coordinate_points}
\title{Produce the model circle points.}
\usage{
produce_model_coordinate_points(
  from = 0,
  by = 5,
  to = 90,
  break_points = NULL,
  radius = 1
)
}
\arguments{
\item{from}{: from angle in degree}

\item{by}{: by angle in degree}

\item{to}{: to angle in degree}

\item{break_points}{the angles points in degree}

\item{radius}{the radius of circle}
}
\value{
matrix, see details
}
\description{
It has many applications:
\enumerate{
\item produce poly regular shapes: like triangles, rectangular and so on.
\item just drow a circle, with points.
}
}
\details{
The return values formats are:
a 2 x n matrix, first row is x axis, second row is y axis.
}
\examples{
produce_model_coordinate_points()
}
