% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class_PathwayAPI.R
\name{pScaleCoord}
\alias{pScaleCoord}
\title{Assign the Scale Coordinate in this function.}
\usage{
pScaleCoord(oriWidth = 1, oriHeight = 1, destWidth = 1, destHeight = 1)
}
\arguments{
\item{realWidth}{the width that you scale to}

\item{realHeight}{the height that you scale to}
}
\value{
the \code{pDescriptor} instance
}
\description{
Assign the Scale Coordinate in this function.
}
\details{
For example, if users assign a rectangle with \verb{x = 0.5, y = 0.5, w = 0.5, h =0.5} ,
while the \verb{width = 1, height = 1}, and assign the \code{pScaleCoordinate(2, 2)} .
Then this function can scale the coordinate, making the
rectangle as \verb{x = 1, y = 1, w = 1, h = 1}.
}
\examples{
pScaleCoord(2, 2)
}
