% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class_PathwayAPI.R
\name{pTitle}
\alias{pTitle}
\title{Assign the title in this function.}
\usage{
pTitle(name, gpar_text = NULL)
}
\arguments{
\item{name}{the title string}

\item{gpar_text}{the gprahic parameter of the text}
}
\value{
the \code{pDescriptor} instance
}
\description{
Assign the title in this function.
}
\examples{
pTitle("This is title")
}
