% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Global_variables.R
\name{register_global_bioGraphics_nodes_list}
\alias{register_global_bioGraphics_nodes_list}
\title{Register global bioGraphics nodes list}
\usage{
register_global_bioGraphics_nodes_list(name, value)
}
\arguments{
\item{name}{your grob name}

\item{value}{a grob}
}
\description{
Register global bioGraphics nodes list
}
\examples{
register_global_bioGraphics_nodes_list('name', grob)
}
