% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class_MembraneDNA_AlongLineDrawer.R
\name{RoundRectangularBiLayerDrawer}
\alias{RoundRectangularBiLayerDrawer}
\title{The round rectangule model drawer, it can draw rectangule along the curve.}
\description{
The draw of the rectangule is repeated, so it should has a drawer class.
}
\examples{

## ------------------------------------------------
## Method `RoundRectangularBiLayerDrawer$calculate_along_curve`
## ------------------------------------------------

calculate_along_curve(1:5,1:5)
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{numOfIntervals}}{the number of intervals, default is 10}

\item{\code{percentageOfRectangle}}{The percentage of the round rectangle, greater 0 and less than 1. Default is 0.7}

\item{\code{roundRectGpar}}{The graphics par of the round rectangle}

\item{\code{desired_height}}{The height in inch's}

\item{\code{painting_parameters}}{store for the parameters for painting. (for more complex functions it can be used)}

\item{\code{n_pointToSmoothOut}}{see \code{\link[=soomthOut_twoSidesOf_quadrangle]{soomthOut_twoSidesOf_quadrangle()}}}

\item{\code{n_step_bezier}}{see \code{\link[=soomthOut_twoSidesOf_quadrangle]{soomthOut_twoSidesOf_quadrangle()}}
calculate lipid along curve, return the needed values for ploting.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-RoundRectangularBiLayerDrawer-calculate_along_curve}{\code{RoundRectangularBiLayerDrawer$calculate_along_curve()}}
\item \href{#method-RoundRectangularBiLayerDrawer-draw_along_curve}{\code{RoundRectangularBiLayerDrawer$draw_along_curve()}}
\item \href{#method-RoundRectangularBiLayerDrawer-clone}{\code{RoundRectangularBiLayerDrawer$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RoundRectangularBiLayerDrawer-calculate_along_curve"></a>}}
\if{latex}{\out{\hypertarget{method-RoundRectangularBiLayerDrawer-calculate_along_curve}{}}}
\subsection{Method \code{calculate_along_curve()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RoundRectangularBiLayerDrawer$calculate_along_curve(lineX, lineY)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{lineX}}{the x axis points}

\item{\code{lineY}}{the y axis points}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the list of calculated values
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{calculate_along_curve(1:5,1:5)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RoundRectangularBiLayerDrawer-draw_along_curve"></a>}}
\if{latex}{\out{\hypertarget{method-RoundRectangularBiLayerDrawer-draw_along_curve}{}}}
\subsection{Method \code{draw_along_curve()}}{
Input the curve line on the x,y parameters, the lipid will correctly draw along the line.


It will handle the angle automatically. Note the denser the dots, the more they are drawn
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RoundRectangularBiLayerDrawer$draw_along_curve(lineX, lineY)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{lineX}}{x values}

\item{\code{lineY}}{y values}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RoundRectangularBiLayerDrawer-clone"></a>}}
\if{latex}{\out{\hypertarget{method-RoundRectangularBiLayerDrawer-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RoundRectangularBiLayerDrawer$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
