% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Global_variables.R
\name{set_global_text_pars}
\alias{set_global_text_pars}
\title{Quick set the text graphic parameters.}
\usage{
set_global_text_pars(name, value)
}
\arguments{
\item{name}{the name of the parameter}

\item{value}{the value}
}
\description{
all the \code{\link[=gpar]{gpar()}} function supported parameter names, like:\if{html}{\out{
<pre>
col	Colour for lines and borders.
fill	Colour for filling rectangles, polygons, ...
alpha	Alpha channel for transparency
lty	Line type
lwd	Line width
lex	Multiplier applied to line width
lineend	Line end style (round, butt, square)
linejoin	Line join style (round, mitre, bevel)
linemitre	Line mitre limit (number greater than 1)
fontsize	The size of text (in points)
cex	Multiplier applied to fontsize
fontfamily	The font family
fontface	The font face (bold, italic, ...)
lineheight	The height of a line as a multiple of the size of text
font	Font face (alias for fontface; for backward compatibility)
</pre>
}}
}
\examples{
set_global_text_pars('fontsize',  7)
}
