% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Display.R
\name{show_bioGraphicNodes_models}
\alias{show_bioGraphicNodes_models}
\title{A snapshot of the bioGraphics nodes.}
\usage{
show_bioGraphicNodes_models()
}
\description{
Also, we support the free shape nodel.
}
\examples{
show_bioGraphicNodes_models()
}
