% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Display.R
\name{show_grid_forAssistance}
\alias{show_grid_forAssistance}
\title{Generate the grid for assistant.}
\usage{
show_grid_forAssistance(interval = 1)
}
\arguments{
\item{interval}{the gap, in inch unit}
}
\description{
Most of the time, we need to use the coordinates to locate the complex.
}
\examples{
show_grid_forAssistance(0.5)
}
