% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Display.R
\name{show_pathway_models}
\alias{show_pathway_models}
\title{Display a basic signalling pathway model.}
\usage{
show_pathway_models()
}
\description{
Display a basic signalling pathway model.
}
\examples{
show_pathway_models()
}
