% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Display.R
\name{show_phospholipid_models}
\alias{show_phospholipid_models}
\title{Display the supported phospholipid models.}
\usage{
show_phospholipid_models()
}
\description{
Display the supported phospholipid models.
}
\examples{
show_phospholipid_models()
}
