% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Display.R
\name{show_roundRect_biLayer}
\alias{show_roundRect_biLayer}
\title{Display the round rectangle bi-layer phospholipid model}
\usage{
show_roundRect_biLayer()
}
\description{
Display the round rectangle bi-layer phospholipid model
}
\examples{
show_roundRect_biLayer()
}
