% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class_AffineTransformer.R
\name{simulate_lipid_bilayer_model}
\alias{simulate_lipid_bilayer_model}
\title{Simulate the lipid bilayer model.}
\usage{
simulate_lipid_bilayer_model()
}
\value{
a list
}
\description{
Simulate the lipid bilayer model.
}
\examples{
simulate_lipid_bilayer_model()
}
