% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class_Geometric.R
\name{soomthOut_twoSidesOf_quadrangle}
\alias{soomthOut_twoSidesOf_quadrangle}
\title{Soomth out the left and right sides of the quadrangle.}
\usage{
soomthOut_twoSidesOf_quadrangle(df, n_pointToSmoothOut = 2, n_step_bezier = 40)
}
\arguments{
\item{df}{the data.frame or matrix of the x y points, see details}

\item{n_pointToSmoothOut}{number of points to used to smooth out}

\item{n_step_bezier}{number of steps of the bezier}
}
\value{
a list with components x and y
}
\description{
Soomth out the left and right sides of the quadrangle.
}
\details{
the \code{df} parameters: first row is the top side line x, second row is the top side line y.
The third row is the bottom line x and the fourth row is the bottom line y.
}
\examples{
grid.newpage()
xx <- seq.int(from = 1, to = 4, by = 0.2)
df <- rbind(xx,1, xx, 2)
# grid.lines(c(xx,rev(xx)), c(df[2,], rev(df[4,])), default.units = 'in')
tt <- soomthOut_twoSidesOf_quadrangle(df,n_pointToSmoothOut = 1 , n_step_bezier = 40)
tt <- soomthOut_twoSidesOf_quadrangle(df,n_pointToSmoothOut = 2 , n_step_bezier = 40)
tt <- soomthOut_twoSidesOf_quadrangle(df,n_pointToSmoothOut = 3 , n_step_bezier = 40)
tt <- soomthOut_twoSidesOf_quadrangle(df,n_pointToSmoothOut = 4 , n_step_bezier = 40)
tt <- soomthOut_twoSidesOf_quadrangle(df,n_pointToSmoothOut = 5 , n_step_bezier = 50)
grid.polygon(tt[1,], tt[2,], default.units = 'in')
}
