% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{basic_bioGraphics_templates}
\alias{basic_bioGraphics_templates}
\title{Create the basic bio graphics node templates
Besides the circle and ellipse shape}
\format{
An object of class \code{list} of length 2.
}
\usage{
basic_bioGraphics_templates
}
\description{
Create the basic bio graphics node templates
Besides the circle and ellipse shape
}
\keyword{datasets}
