% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class_BioGraphicsObjs.R
\name{BioGraphicNode}
\alias{BioGraphicNode}
\title{The BioGraphicNode model drawer}
\description{
\code{BioGraphicNode} is an models for user to draw any arbitrary node

It can be organized by list in list structure. Some situations are the most suitable:
\itemize{
\item The complex Membrane receptor protein.
\item Some complex structure, like DC complex in WNT signaling pathway.
}
}
\examples{

## ------------------------------------------------
## Method `BioGraphicNode$draw_me`
## ------------------------------------------------

draw_me()
The print method
Override here
}
\section{Super classes}{
\code{\link[PathwayIlluminator:ListTreeNode]{PathwayIlluminator::ListTreeNode}} -> \code{\link[PathwayIlluminator:GraphicNode]{PathwayIlluminator::GraphicNode}} -> \code{BioGraphicNode}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{circle_radius}}{how big the node is.}

\item{\code{rotation_angle_inRadian}}{the rotation angle in radian, default is 0.}

\item{\code{xyPoints_shape}}{the x y points  of the drawing shape, default is a rectangle}

\item{\code{gpar_shape}}{the gpar for the shape}

\item{\code{gpar_text}}{the gpar for the text}

\item{\code{inner_extension_ratio}}{the ratio of inner boarder, try this parameter by yourself.}

\item{\code{text_position}}{hjust and vjust for textGrob
Draw if this node does not have children.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-BioGraphicNode-draw_if_hasChildren}{\code{BioGraphicNode$draw_if_hasChildren()}}
\item \href{#method-BioGraphicNode-draw_if_noChildren}{\code{BioGraphicNode$draw_if_noChildren()}}
\item \href{#method-BioGraphicNode-draw_me}{\code{BioGraphicNode$draw_me()}}
\item \href{#method-BioGraphicNode-print}{\code{BioGraphicNode$print()}}
\item \href{#method-BioGraphicNode-clone}{\code{BioGraphicNode$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="PathwayIlluminator" data-topic="ListTreeNode" data-id="addChild"><a href='../../PathwayIlluminator/html/ListTreeNode.html#method-ListTreeNode-addChild'><code>PathwayIlluminator::ListTreeNode$addChild()</code></a></span></li>
<li><span class="pkg-link" data-pkg="PathwayIlluminator" data-topic="ListTreeNode" data-id="getIDName"><a href='../../PathwayIlluminator/html/ListTreeNode.html#method-ListTreeNode-getIDName'><code>PathwayIlluminator::ListTreeNode$getIDName()</code></a></span></li>
<li><span class="pkg-link" data-pkg="PathwayIlluminator" data-topic="ListTreeNode" data-id="initialize"><a href='../../PathwayIlluminator/html/ListTreeNode.html#method-ListTreeNode-initialize'><code>PathwayIlluminator::ListTreeNode$initialize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="PathwayIlluminator" data-topic="ListTreeNode" data-id="isLeaf"><a href='../../PathwayIlluminator/html/ListTreeNode.html#method-ListTreeNode-isLeaf'><code>PathwayIlluminator::ListTreeNode$isLeaf()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BioGraphicNode-draw_if_hasChildren"></a>}}
\if{latex}{\out{\hypertarget{method-BioGraphicNode-draw_if_hasChildren}{}}}
\subsection{Method \code{draw_if_hasChildren()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BioGraphicNode$draw_if_hasChildren(default_unit = "in")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{default_unit}}{the unit default is inch.
Draw if this node does have children.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BioGraphicNode-draw_if_noChildren"></a>}}
\if{latex}{\out{\hypertarget{method-BioGraphicNode-draw_if_noChildren}{}}}
\subsection{Method \code{draw_if_noChildren()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BioGraphicNode$draw_if_noChildren(default_unit = "in")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{default_unit}}{the unit default is inch.
Draw myself}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BioGraphicNode-draw_me"></a>}}
\if{latex}{\out{\hypertarget{method-BioGraphicNode-draw_me}{}}}
\subsection{Method \code{draw_me()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BioGraphicNode$draw_me(default_unit = "in")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{default_unit}}{the coordinate unit, default is 'in'}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{draw_me()
The print method
Override here
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BioGraphicNode-print"></a>}}
\if{latex}{\out{\hypertarget{method-BioGraphicNode-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BioGraphicNode$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BioGraphicNode-clone"></a>}}
\if{latex}{\out{\hypertarget{method-BioGraphicNode-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BioGraphicNode$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
