% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class_Geometric.R
\name{do_bilateral_extension_alongCurve}
\alias{do_bilateral_extension_alongCurve}
\title{Employ the bilateral extansion along a curve}
\usage{
do_bilateral_extension_alongCurve(x, y, d = 1)
}
\arguments{
\item{x}{x axis of curve}

\item{y}{y axis of curve}

\item{d}{Euclidean distance}
}
\value{
a list of two line points
}
\description{
Employ the bilateral extansion along a curve
}
