% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class_AffineTransformer.R
\name{do_scale_rotate_translate_affineTransfor}
\alias{do_scale_rotate_translate_affineTransfor}
\title{First scale, next rotate and finally translate.}
\usage{
do_scale_rotate_translate_affineTransfor(
  mat,
  scaleWidth = 1,
  scaleHeight = 1,
  theta,
  moveX,
  moveY
)
}
\arguments{
\item{mat}{the input matrix}

\item{scaleWidth}{scale of width}

\item{scaleHeight}{scale of height}

\item{theta}{angle in radian}

\item{moveX}{translate x}

\item{moveY}{translate y}
}
\value{
the transformed matrix
}
\description{
First scale, next rotate and finally translate.
}
\examples{
do_scale_rotate_translate_affineTransfor( your requirement)
}
