% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Global_variables.R
\name{get_global_bioGraphics_nodes_list}
\alias{get_global_bioGraphics_nodes_list}
\title{Each bioGraphicsNode instance as well as some properties will be record, so users can get the instance here.}
\usage{
get_global_bioGraphics_nodes_list()
}
\value{
a list of the needed information
}
\description{
The record will increase by the [\code{\link[=register_global_bioGraphics_nodes_list]{register_global_bioGraphics_nodes_list()}}] function.
}
\details{
The returned list is consists of the elements, which has the property of \code{width} and \code{height}, as well as a \code{grob}.
The \code{width} and \code{height} are the center of the \code{grob}.
}
\examples{
get_global_bioGraphics_nodes_list() |> names()
}
