% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class_LipidBilayerOnLine.R
\name{get_rotation_of_twoPoints}
\alias{get_rotation_of_twoPoints}
\title{Get the rotation of the two points.}
\usage{
get_rotation_of_twoPoints(x1, y1, x2, y2)
}
\arguments{
\item{y2}{}
}
\value{
the rotation angle in radian
}
\description{
Get the rotation of the two points.
}
\examples{
get_rotation_of_twoPoints(0,0,1,1)
}
