% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lipid_models}
\alias{lipid_models}
\title{Create the lipid models, the model is stored in this.}
\format{
a list in list structure
}
\source{
manuly create my developers.
A data frame with 7,240 rows and 60 columns:
\describe{
\item{country}{Country name}
\item{iso2, iso3}{2 & 3 letter ISO country codes}
\item{year}{Year}
...
}
}
\usage{
lipid_models
}
\description{
Create the lipid models, the model is stored in this.
}
\keyword{datasets}
