% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Util_graphic_common.R
\name{polar2cartesianCoor}
\alias{polar2cartesianCoor}
\title{Polar coordinates to Cartesian coordinates.}
\usage{
polar2cartesianCoor(radius, angle, xCenter = 0.5, yCenter = 0.5)
}
\arguments{
\item{radius:}{the circular radius}

\item{angle:}{unit is in degrees, NOT radian.}

\item{xCenter:}{the xCenter coordinate point}

\item{yCenter:}{the yCenter coordinate point}
}
\value{
: first column is the sin related values, second column is cos related values.
}
\description{
The function supports the vectorised invoke usage.
The implementation employs the euler's formula to accelerate the computation.
NOTE the radius and location (xCenter, yCenter) should be the same unit with values(NOT unit instances).
}
\examples{
polar2cartesianCoor(1,5)

}
