% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Global_variables.R
\name{set_global_pars}
\alias{set_global_pars}
\title{Set the parameters for the global value.}
\usage{
set_global_pars(name, value)
}
\arguments{
\item{name}{e.g. blank_area_ratio}

\item{value}{the value could be list or single value or automic vector}
}
\description{
Set the parameters for the global value.
}
\examples{
set_global_pars('blank_area_ratio', list(l = 0.05, r = 0.05, t = 0.02, b = 0.02))
}
